fix: consolidate pkg/api and pkg/config files for file-count CI
Merge factory validation API tests and canonical structural validation
helpers into existing package files so pkg-file-count limits pass.
