+++ pkg/api/server_factory_test.go
@@ -514,3 +514,122 @@ func TestLegacyCreateFactoryRoute_RemovedFromRouter(t *testing.T) {
 		t.Fatalf("POST /factory status = %d, want route removed", rec.Code)
 	}
 }
+
+func TestValidateFactory_ReturnsEmptyTargetsForValidFactory(t *testing.T) {
+	srv := newTestServer(&testutil.MockFactory{})
+
+	req := httptest.NewRequest(http.MethodPost, "/factory-validations", bytes.NewBufferString(validNamedFactoryBody("beta", "beta-task")))
+	req.Header.Set("Content-Type", "application/json")
+	rec := httptest.NewRecorder()
+	srv.Handler().ServeHTTP(rec, req)
+
+	if rec.Code != http.StatusOK {
+		t.Fatalf("expected 200, got %d: %s", rec.Code, rec.Body.String())
+	}
+
+	result := decodeJSONResponse[factoryapi.FactoryValidationResult](t, rec)
+	if len(result.Targets) != 0 {
+		t.Fatalf("targets = %#v, want empty slice", result.Targets)
+	}
+}
+
+func TestValidateFactory_ReturnsMultipleTargetsForInvalidFactory(t *testing.T) {
+	srv := newTestServer(&testutil.MockFactory{})
+
+	body := `{
+		"name":"alpha",
+		"workTypes":[{"name":"story","states":[
+			{"name":"queued","type":"INITIAL"},
+			{"name":"queued-dup","type":"PROCESSING"}
+		]}],
+		"workers":[{"name":"worker-a"},{"name":"worker-a"}],
+		"workstations":[{
+			"name":"process",
+			"worker":"missing-worker",
+			"inputs":[{"workType":"story","state":"queued"}],
+			"outputs":[{"workType":"story","state":"missing-state"}]
+		}]
+	}`
+
+	req := httptest.NewRequest(http.MethodPost, "/factory-validations", bytes.NewBufferString(body))
+	req.Header.Set("Content-Type", "application/json")
+	rec := httptest.NewRecorder()
+	srv.Handler().ServeHTTP(rec, req)
+
+	if rec.Code != http.StatusOK {
+		t.Fatalf("expected 200, got %d: %s", rec.Code, rec.Body.String())
+	}
+
+	result := decodeJSONResponse[factoryapi.FactoryValidationResult](t, rec)
+	if len(result.Targets) < 2 {
+		t.Fatalf("targets = %d, want multiple validation targets", len(result.Targets))
+	}
+	assertHasValidationTargetCode(t, result.Targets, factoryvalidation.CodeDuplicateIdentifier)
+	assertHasValidationTargetCode(t, result.Targets, factoryvalidation.CodeDanglingWorkerReference)
+	assertHasValidationTargetCode(t, result.Targets, factoryvalidation.CodeDanglingPlaceReference)
+}
+
+func TestValidateFactory_RejectsMalformedPayload(t *testing.T) {
+	srv := newTestServer(&testutil.MockFactory{})
+
+	req := httptest.NewRequest(http.MethodPost, "/factory-validations", bytes.NewBufferString(`{"name":"alpha"`))
+	req.Header.Set("Content-Type", "application/json")
+	rec := httptest.NewRecorder()
+	srv.Handler().ServeHTTP(rec, req)
+
+	if rec.Code != http.StatusBadRequest {
+		t.Fatalf("expected 400, got %d: %s", rec.Code, rec.Body.String())
+	}
+}
+
+func TestSaveCurrentFactory_ReturnsBobWorkstationOnFailureTarget(t *testing.T) {
+	target := factoryapi.FactoryValidationTarget{
+		Code:     factoryvalidation.CodeWorkstationMissingFailureRoute,
+		Severity: factoryapi.FactoryValidationSeverityError,
+		Message:  `workstation "bob" must define a failure route.`,
+		Subject: factoryapi.FactoryValidationSubject{
+			Type:     factoryapi.FactoryValidationSubjectTypeWorkstation,
+			Id:       "bob",
+			Location: factoryapi.FactoryValidationSubjectLocationOnFailure,
+		},
+	}
+	srv := newTestServer(&testutil.MockFactory{
+		SaveCurrentFactoryErr: apisurface.NewTopologyValidationError(
+			"Factory topology contains invalid graph references.",
+			[]factoryapi.FactoryValidationTarget{target},
+		),
+	})
+
+	req := httptest.NewRequest(http.MethodPut, "/factory-sessions/~default/factory", bytes.NewBufferString(`{"name":"beta"}`))
+	req.Header.Set("Content-Type", "application/json")
+	rec := httptest.NewRecorder()
+	srv.Handler().ServeHTTP(rec, req)
+
+	response := decodeJSONResponse[factoryapi.ErrorResponse](t, rec)
+	if rec.Code != http.StatusBadRequest || response.Code != factoryapi.INVALIDFACTORY {
+		t.Fatalf("response = %#v status=%d", response, rec.Code)
+	}
+	if response.Targets == nil || len(*response.Targets) != 1 {
+		t.Fatalf("targets = %#v, want one canonical target", response.Targets)
+	}
+	got := (*response.Targets)[0]
+	assertHasValidationTarget(
+		t,
+		[]factoryapi.FactoryValidationTarget{got},
+		factoryvalidation.CodeWorkstationMissingFailureRoute,
+		factoryapi.FactoryValidationSubjectTypeWorkstation,
+		"bob",
+		factoryapi.FactoryValidationSubjectLocationOnFailure,
+		"bob ON_FAILURE target",
+	)
+}
+
+func assertHasValidationTargetCode(t *testing.T, targets []factoryapi.FactoryValidationTarget, code string) {
+	t.Helper()
+	for _, target := range targets {
+		if target.Code == code {
+			return
+		}
+	}
+	t.Fatalf("targets = %#v, want code %q", targets, code)
+}
