--- pkg/api/server_factory_validation_test.go
@@ -1,89 +1,0 @@
-package api
-
-import (
-	"bytes"
-	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
-	factoryvalidation "github.com/portpowered/infinite-you/pkg/factory/validation"
-	"github.com/portpowered/infinite-you/pkg/testutil"
-)
-
-func TestValidateFactory_ReturnsEmptyTargetsForValidFactory(t *testing.T) {
-	srv := newTestServer(&testutil.MockFactory{})
-
-	req := httptest.NewRequest(http.MethodPost, "/factory-validations", bytes.NewBufferString(validNamedFactoryBody("beta", "beta-task")))
-	req.Header.Set("Content-Type", "application/json")
-	rec := httptest.NewRecorder()
-	srv.Handler().ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Fatalf("expected 200, got %d: %s", rec.Code, rec.Body.String())
-	}
-
-	result := decodeJSONResponse[factoryapi.FactoryValidationResult](t, rec)
-	if len(result.Targets) != 0 {
-		t.Fatalf("targets = %#v, want empty slice", result.Targets)
-	}
-}
-
-func TestValidateFactory_ReturnsMultipleTargetsForInvalidFactory(t *testing.T) {
-	srv := newTestServer(&testutil.MockFactory{})
-
-	body := `{
-		"name":"alpha",
-		"workTypes":[{"name":"story","states":[
-			{"name":"queued","type":"INITIAL"},
-			{"name":"queued-dup","type":"PROCESSING"}
-		]}],
-		"workers":[{"name":"worker-a"},{"name":"worker-a"}],
-		"workstations":[{
-			"name":"process",
-			"worker":"missing-worker",
-			"inputs":[{"workType":"story","state":"queued"}],
-			"outputs":[{"workType":"story","state":"missing-state"}]
-		}]
-	}`
-
-	req := httptest.NewRequest(http.MethodPost, "/factory-validations", bytes.NewBufferString(body))
-	req.Header.Set("Content-Type", "application/json")
-	rec := httptest.NewRecorder()
-	srv.Handler().ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusOK {
-		t.Fatalf("expected 200, got %d: %s", rec.Code, rec.Body.String())
-	}
-
-	result := decodeJSONResponse[factoryapi.FactoryValidationResult](t, rec)
-	if len(result.Targets) < 2 {
-		t.Fatalf("targets = %d, want multiple validation targets", len(result.Targets))
-	}
-	assertHasValidationTargetCode(t, result.Targets, factoryvalidation.CodeDuplicateIdentifier)
-	assertHasValidationTargetCode(t, result.Targets, factoryvalidation.CodeDanglingWorkerReference)
-	assertHasValidationTargetCode(t, result.Targets, factoryvalidation.CodeDanglingPlaceReference)
-}
-
-func TestValidateFactory_RejectsMalformedPayload(t *testing.T) {
-	srv := newTestServer(&testutil.MockFactory{})
-
-	req := httptest.NewRequest(http.MethodPost, "/factory-validations", bytes.NewBufferString(`{"name":"alpha"`))
-	req.Header.Set("Content-Type", "application/json")
-	rec := httptest.NewRecorder()
-	srv.Handler().ServeHTTP(rec, req)
-
-	if rec.Code != http.StatusBadRequest {
-		t.Fatalf("expected 400, got %d: %s", rec.Code, rec.Body.String())
-	}
-}
-
-func assertHasValidationTargetCode(t *testing.T, targets []factoryapi.FactoryValidationTarget, code string) {
-	t.Helper()
-	for _, target := range targets {
-		if target.Code == code {
-			return
-		}
-	}
-	t.Fatalf("targets = %#v, want code %q", targets, code)
-}
