+++ pkg/api/server_test_helpers_test.go
@@ -22,6 +22,28 @@ import (
 	"go.uber.org/zap"
 )
 
+func assertHasValidationTarget(
+	t *testing.T,
+	targets []factoryapi.FactoryValidationTarget,
+	code string,
+	subjectType factoryapi.FactoryValidationSubjectType,
+	subjectID string,
+	location factoryapi.FactoryValidationSubjectLocation,
+	want string,
+) {
+	t.Helper()
+	for _, target := range targets {
+		if target.Code != code {
+			continue
+		}
+		if target.Subject.Type != subjectType || target.Subject.Id != subjectID || target.Subject.Location != location {
+			continue
+		}
+		return
+	}
+	t.Fatalf("validation targets = %#v, want %s", targets, want)
+}
+
 func newTestServer(f *testutil.MockFactory) *Server {
 	logger, _ := zap.NewDevelopment()
 	return NewServer(f, 8080, logger)
