--- pkg/api/validation_target_test_helpers.go
@@ -1,29 +1,0 @@
-package api
-
-import (
-	"testing"
-
-	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
-)
-
-func assertHasValidationTarget(
-	t *testing.T,
-	targets []factoryapi.FactoryValidationTarget,
-	code string,
-	subjectType factoryapi.FactoryValidationSubjectType,
-	subjectID string,
-	location factoryapi.FactoryValidationSubjectLocation,
-	want string,
-) {
-	t.Helper()
-	for _, target := range targets {
-		if target.Code != code {
-			continue
-		}
-		if target.Subject.Type != subjectType || target.Subject.Id != subjectID || target.Subject.Location != location {
-			continue
-		}
-		return
-	}
-	t.Fatalf("validation targets = %#v, want %s", targets, want)
-}
